+++ src/delfino/main.py
@@ -2,20 +2,20 @@ import logging
 import os
 import sys
 from pathlib import Path
-from typing import Any, Dict, List, Optional, Set, Union
+from typing import Any, List, Optional, Union
+from warnings import warn
 
 import click
 import toml
 from pydantic import ValidationError
 
-from delfino import commands
-from delfino.click_utils.command import find_commands
+from delfino.click_utils.command import CommandRegistry
 from delfino.click_utils.completion import install_completion_option, show_completion_option
 from delfino.click_utils.help import extended_help_option
 from delfino.click_utils.verbosity import log_level_option
-from delfino.constants import COMMANDS_DIRECTORY_NAME, ENTRY_POINT, PYPROJECT_TOML_FILENAME
+from delfino.constants import ENTRY_POINT, PYPROJECT_TOML_FILENAME
 from delfino.contexts import AppContext
-from delfino.models.pyproject_toml import PyprojectToml
+from delfino.models.pyproject_toml import PluginConfig, PyprojectToml
 from delfino.utils import get_package_manager
 
 
@@ -31,31 +31,39 @@ class Commands(click.MultiCommand):
         # When evaluating available commands, the program should not fail. We save the exception
         # to show it only when a command is executed.
         self._pyproject_toml: Union[PyprojectToml, Exception]
-        self._hidden_plugins: Set[str] = set()
 
         try:
             self._pyproject_toml = PyprojectToml(file_path=pyproject_toml_path, **toml.load(pyproject_toml_path))
-            self._hidden_plugins = self._pyproject_toml.tool.delfino.disable_commands
+
+            plugins = self._pyproject_toml.tool.delfino.plugins
+
+            # Temporary code to deprecate tool.delfino.disable_commands
+            if self._pyproject_toml.tool.delfino.disable_commands:
+                plugins[CommandRegistry.CORE_PLUGIN_NAME] = PluginConfig(
+                    disable_commands=self._pyproject_toml.tool.delfino.disable_commands
+                )
+                warn(
+                    f"```\n[tool.delfino]\ndisable_commands = ...\n```\nconfiguration option "
+                    f"in `pyproject.toml` is deprecated and will be removed in the future. Please use\n"
+                    f"```\n[tool.delfino.plugins.{CommandRegistry.CORE_PLUGIN_NAME}]\ndisable_commands = ...\n```\n"
+                    f"instead.",
+                    DeprecationWarning,
+                )
         except ValidationError as exc:
             self._pyproject_toml = exc
         except FileNotFoundError:
             self._pyproject_toml = PyprojectToml()
 
-        self._plugins: Dict[str, click.Command] = find_commands(commands.__package__, required=True)
-        self._plugins.update(find_commands(COMMANDS_DIRECTORY_NAME, required=False))
-        self._plugins.update(find_commands("tasks", required=False, new_name=COMMANDS_DIRECTORY_NAME))
-
-        for name, cmd in list(self._plugins.items()):
-            self._plugins[name] = extended_help_option(cmd)
+        self._command_registry = CommandRegistry(plugins)
 
     def list_commands(self, ctx: click.Context) -> List[str]:
-        """Override to hide commands marked as hidden in the ``pyproject.toml`` file."""
+        """Override as MultiCommand always returns []."""
         del ctx
-        return sorted(set(self._plugins.keys()).difference(self._hidden_plugins))
+        return sorted(command.name for command in self._command_registry.visible_commands)
 
     def get_command(self, ctx: click.Context, cmd_name: str) -> Optional[click.Command]:
         """Override to give all commands a common ``AppContext`` or fail if ``pyproject.toml`` is broken/missing."""
-        cmd = self._plugins.get(cmd_name, None)
+        cmd = self._command_registry.get(cmd_name, None)
 
         if ctx.resilient_parsing:  # do not fail on auto-completion
             return cmd
@@ -83,12 +91,19 @@ class Commands(click.MultiCommand):
     def get_help(self, *args, **kwargs) -> str:
         help_str = super().get_help(*args, **kwargs)
 
-        if self._hidden_plugins:
+        if self._command_registry.hidden_commands:
             if logging.root.level == logging.DEBUG:
                 help_str += click.style(
-                    f"\n\nDisabled command{'s' if len(self._hidden_plugins) > 1 else ''}: "
-                    + ", ".join(sorted(self._hidden_plugins))
-                    + f" (see 'tool.{ENTRY_POINT}.disable_commands' in '{PYPROJECT_TOML_FILENAME}')",
+                    f"\n\nDisabled command{'s' if len(self._command_registry.hidden_commands) > 1 else ''}: "
+                    + ", ".join(
+                        sorted(
+                            [
+                                f"{command.plugin_name}/{command.command.name}"
+                                for command in self._command_registry.hidden_commands
+                            ]
+                        )
+                    )
+                    + f" (see 'tool.{ENTRY_POINT}.disable_plugin_commands' in '{PYPROJECT_TOML_FILENAME}')",
                     fg="white",
                 )