+++ tests/integration/fake_plugins/fake_plugin_a/fake_plugin_a/commands/lint.py
@@ -1,0 +1,11 @@
+import click
+
+
+@click.command()
+def lint():
+    pass
+
+
+@click.command()
+def build():
+    pass