+++ tests/integration/fake_plugins/fake_plugin_without_entry_point/fake_plugin_without_entry_point/commands/__init__.py
@@ -1,0 +1,6 @@
+import click
+
+
+@click.command()
+def no_entry_point_command():
+    pass